Iteration 8

+++ db/migrate/20150929225838_create_rides.rb
@@ -8,7 +8,7 @@ class CreateRides < ActiveRecord::Migration
       t.datetime :accepted_time
       t.datetime :pickup_time
       t.datetime :dropoff_time
-      t.integer :cost
+      t.float :cost
       t.integer :rider_id
       t.integer :driver_id
       t.timestamps null: false
